commiting civi and drupal module updates

+++ sites/all/modules/contrib/civicrm/CRM/Core/BAO/SchemaHandler.php
@@ -1,9 +1,9 @@
 <?php
 /*
  +--------------------------------------------------------------------+
- | CiviCRM version 4.5                                                |
+ | CiviCRM version 4.7                                                |
  +--------------------------------------------------------------------+
- | Copyright CiviCRM LLC (c) 2004-2014                                |
+ | Copyright CiviCRM LLC (c) 2004-2017                                |
  +--------------------------------------------------------------------+
  | This file is a part of CiviCRM.                                    |
  |                                                                    |
@@ -23,58 +23,48 @@
  | GNU Affero General Public License or the licensing of CiviCRM,     |
  | see the CiviCRM license FAQ at http://civicrm.org/licensing        |
  +--------------------------------------------------------------------+
-*/
+ */
 
 /**
  *
  * @package CRM
- * @copyright CiviCRM LLC (c) 2004-2014
- * $Id$
- *
+ * @copyright CiviCRM LLC (c) 2004-2017
  */
 
 /**
- *  This file contains functions for creating and altering CiviCRM-tables.
- */
-
-/**
- * structure, similar to what is used in GenCode.php
+ *  This file contains functions for creating and altering CiviCRM-tables structure.
  *
  * $table = array(
-    'name'       => TABLE_NAME,
- *                'attributes' => ATTRIBUTES,
- *                'fields'     => array(
- *                                      array(
-    'name'          => FIELD_NAME,
- *                                             'type'          => FIELD_SQL_TYPE,
- // can be field, index, constraint
- *                                             'class'         => FIELD_CLASS_TYPE,
- *                                             'primary'       => BOOLEAN,
- *                                             'required'      => BOOLEAN,
- *                                             'searchable'    => true,
- *                                             'fk_table_name' => FOREIGN_KEY_TABLE_NAME,
- *                                             'fk_field_name' => FOREIGN_KEY_FIELD_NAME,
- *                                             'comment'       => COMMENT,
- *                                             'default'       => DEFAULT, )
- *                                      ...
- *                                      ) );
+ *  'name'  => TABLE_NAME,
+ *  'attributes' => ATTRIBUTES,
+ *  'fields' => array(
+ *    array(
+ *      'name' => FIELD_NAME,
+ *      // can be field, index, constraint
+ *      'type' => FIELD_SQL_TYPE,
+ *      'class'         => FIELD_CLASS_TYPE,
+ *      'primary'       => BOOLEAN,
+ *      'required'      => BOOLEAN,
+ *      'searchable'    => TRUE,
+ *      'fk_table_name' => FOREIGN_KEY_TABLE_NAME,
+ *      'fk_field_name' => FOREIGN_KEY_FIELD_NAME,
+ *      'comment'       => COMMENT,
+ *      'default'       => DEFAULT, )
+ *      ...
+ *  ));
  */
 class CRM_Core_BAO_SchemaHandler {
 
   /**
-   * Function for creating a civiCRM-table
+   * Create a CiviCRM-table
    *
-   * @param $params
+   * @param array $params
    *
-   * @internal param String $tableName name of the table to be created.
-   * @internal param Array $tableAttributes array containing atrributes for the table that needs to be created
+   * @return bool
+   *   TRUE if successfully created, FALSE otherwise
    *
-   * @return true if successfully created, false otherwise
-   *
-   * @static
-   * @access public
    */
-  static function createTable(&$params) {
+  public static function createTable(&$params) {
     $sql = self::buildTableSQL($params);
     // do not i18n-rewrite
     $dao = CRM_Core_DAO::executeQuery($sql, array(), TRUE, NULL, FALSE, FALSE);
@@ -83,8 +73,8 @@ class CRM_Core_BAO_SchemaHandler {
     $config = CRM_Core_Config::singleton();
     if ($config->logging) {
       // logging support
-      $logging = new CRM_Logging_Schema;
-      $logging->fixSchemaDifferencesFor($params['name'], null, FALSE);
+      $logging = new CRM_Logging_Schema();
+      $logging->fixSchemaDifferencesFor($params['name'], NULL, FALSE);
     }
 
     // always do a trigger rebuild for this table
@@ -94,11 +84,11 @@ class CRM_Core_BAO_SchemaHandler {
   }
 
   /**
-   * @param $params
+   * @param array $params
    *
    * @return string
    */
-  static function buildTableSQL(&$params) {
+  public static function buildTableSQL(&$params) {
     $sql = "CREATE TABLE {$params['name']} (";
     if (isset($params['fields']) &&
       is_array($params['fields'])
@@ -129,13 +119,13 @@ class CRM_Core_BAO_SchemaHandler {
   }
 
   /**
-   * @param $params
+   * @param array $params
    * @param $separator
    * @param $prefix
    *
    * @return string
    */
-  static function buildFieldSQL(&$params, $separator, $prefix) {
+  public static function buildFieldSQL(&$params, $separator, $prefix) {
     $sql = '';
     $sql .= $separator;
     $sql .= str_repeat(' ', 8);
@@ -164,13 +154,13 @@ class CRM_Core_BAO_SchemaHandler {
   }
 
   /**
-   * @param $params
+   * @param array $params
    * @param $separator
    * @param $prefix
    *
-   * @return null|string
+   * @return NULL|string
    */
-  static function buildPrimaryKeySQL(&$params, $separator, $prefix) {
+  public static function buildPrimaryKeySQL(&$params, $separator, $prefix) {
     $sql = NULL;
     if (!empty($params['primary'])) {
       $sql .= $separator;
@@ -182,14 +172,14 @@ class CRM_Core_BAO_SchemaHandler {
   }
 
   /**
-   * @param $params
+   * @param array $params
    * @param $separator
    * @param $prefix
    * @param bool $indexExist
    *
-   * @return null|string
+   * @return NULL|string
    */
-  static function buildSearchIndexSQL(&$params, $separator, $prefix, $indexExist = FALSE) {
+  public static function buildSearchIndexSQL(&$params, $separator, $prefix, $indexExist = FALSE) {
     $sql = NULL;
 
     // dont index blob
@@ -215,13 +205,13 @@ class CRM_Core_BAO_SchemaHandler {
   }
 
   /**
-   * @param $params
+   * @param array $params
    * @param $separator
    * @param $prefix
    *
    * @return string
    */
-  static function buildIndexSQL(&$params, $separator, $prefix) {
+  public static function buildIndexSQL(&$params, $separator, $prefix) {
     $sql = '';
     $sql .= $separator;
     $sql .= str_repeat(' ', 8);
@@ -248,12 +238,12 @@ class CRM_Core_BAO_SchemaHandler {
   }
 
   /**
-   * @param $tableName
-   * @param $fkTableName
+   * @param string $tableName
+   * @param string $fkTableName
    *
    * @return bool
    */
-  static function changeFKConstraint($tableName, $fkTableName) {
+  public static function changeFKConstraint($tableName, $fkTableName) {
     $fkName = "{$tableName}_entity_id";
     if (strlen($fkName) >= 48) {
       $fkName = substr($fkName, 0, 32) . "_" . substr(md5($fkName), 0, 16);
@@ -276,14 +266,14 @@ ALTER TABLE {$tableName}
   }
 
   /**
-   * @param $params
+   * @param array $params
    * @param $separator
    * @param $prefix
-   * @param $tableName
+   * @param string $tableName
    *
-   * @return null|string
+   * @return NULL|string
    */
-  static function buildForeignKeySQL(&$params, $separator, $prefix, $tableName) {
+  public static function buildForeignKeySQL(&$params, $separator, $prefix, $tableName) {
     $sql = NULL;
     if (!empty($params['fk_table_name']) && !empty($params['fk_field_name'])) {
       $sql .= $separator;
@@ -301,13 +291,13 @@ ALTER TABLE {$tableName}
   }
 
   /**
-   * @param $params
+   * @param array $params
    * @param bool $indexExist
    * @param bool $triggerRebuild
    *
    * @return bool
    */
-  static function alterFieldSQL(&$params, $indexExist = FALSE, $triggerRebuild = TRUE) {
+  public static function alterFieldSQL(&$params, $indexExist = FALSE, $triggerRebuild = TRUE) {
     $sql = str_repeat(' ', 8);
     $sql .= "ALTER TABLE {$params['table_name']}";
 
@@ -350,14 +340,16 @@ ALTER TABLE {$tableName}
 
     $config = CRM_Core_Config::singleton();
     if ($config->logging) {
-      // logging support: if we’re adding a column (but only then!) make sure the potential relevant log table gets a column as well
-      if ($params['operation'] == 'add') {
-        $logging = new CRM_Logging_Schema;
-        $logging->fixSchemaDifferencesFor($params['table_name'], array('ADD' => array($params['name'])), FALSE);
+      // CRM-16717 not sure why this was originally limited to add.
+      // For example custom tables can have field length changes - which need to flow through to logging.
+      // Are there any modifies we DON'T was to call this function for (& shouldn't it be clever enough to cope?)
+      if ($params['operation'] == 'add' || $params['operation'] == 'modify') {
+        $logging = new CRM_Logging_Schema();
+        $logging->fixSchemaDifferencesFor($params['table_name'], array(trim($prefix) => array($params['name'])), FALSE);
       }
     }
 
-    if($triggerRebuild) {
+    if ($triggerRebuild) {
       CRM_Core_DAO::triggerRebuild($params['table_name']);
     }
 
@@ -365,34 +357,45 @@ ALTER TABLE {$tableName}
   }
 
   /**
-   * Function to delete a civiCRM-table
+   * Delete a CiviCRM-table
    *
-   * @param  String  $tableName   name of the table to be created.
-   *
-   * @return true if successfully deleted, false otherwise
-   *
-   * @static
-   * @access public
+   * @param string $tableName
+   *   Name of the table to be created.
    */
-  static function dropTable($tableName) {
+  public static function dropTable($tableName) {
     $sql = "DROP TABLE $tableName";
     $dao = CRM_Core_DAO::executeQuery($sql);
   }
 
   /**
-   * @param $tableName
-   * @param $columnName
+   * @param string $tableName
+   * @param string $columnName
+   * @param bool $l18n
+   *
    */
-  static function dropColumn($tableName, $columnName) {
-    $sql = "ALTER TABLE $tableName DROP COLUMN $columnName";
-    $dao = CRM_Core_DAO::executeQuery($sql);
+  public static function dropColumn($tableName, $columnName, $l18n = FALSE) {
+    if (self::checkIfFieldExists($tableName, $columnName)) {
+      $sql = "ALTER TABLE $tableName DROP COLUMN $columnName";
+      if ($l18n) {
+        CRM_Core_DAO::executeQuery($sql);
+      }
+      else {
+        CRM_Core_DAO::executeQuery($sql, array(), TRUE, NULL, FALSE, FALSE);
+      }
+      $domain = new CRM_Core_DAO_Domain();
+      $domain->find(TRUE);
+      if ($domain->locales) {
+        $locales = explode(CRM_Core_DAO::VALUE_SEPARATOR, $domain->locales);
+        CRM_Core_I18n_Schema::rebuildMultilingualSchema($locales, NULL);
+      }
+    }
   }
 
   /**
-   * @param $tableName
+   * @param string $tableName
    * @param bool $dropUnique
    */
-  static function changeUniqueToIndex($tableName, $dropUnique = TRUE) {
+  public static function changeUniqueToIndex($tableName, $dropUnique = TRUE) {
     if ($dropUnique) {
       $sql = "ALTER TABLE $tableName
 DROP INDEX `unique_entity_id` ,
@@ -407,16 +410,29 @@ ADD UNIQUE INDEX `unique_entity_id` ( `entity_id` )";
   }
 
   /**
+   * Create indexes.
+   *
    * @param $tables
+   *   Tables to create index for in the format:
+   *     array('civicrm_entity_table' => 'entity_id')
+   *     OR
+   *     array('civicrm_entity_table' => array('entity_id', 'entity_table'))
+   *   The latter will create a combined index on the 2 keys (in order).
+   *
+   *  Side note - when creating combined indexes the one with the most variation
+   *  goes first  - so entity_table always goes after entity_id.
+   *
+   *  It probably makes sense to consider more sophisticated options at some point
+   *  but at the moment this is only being as enhanced as fast as the test is.
+   *
+   * @todo add support for length & multilingual on combined keys.
+   *
    * @param string $createIndexPrefix
    * @param array $substrLenghts
    */
-  static function createIndexes(&$tables, $createIndexPrefix = 'index', $substrLenghts = array(
-    )) {
+  public static function createIndexes($tables, $createIndexPrefix = 'index', $substrLenghts = array()) {
     $queries = array();
-
-        require_once 'CRM/Core/DAO/Domain.php';
-    $domain = new CRM_Core_DAO_Domain;
+    $domain = new CRM_Core_DAO_Domain();
     $domain->find(TRUE);
     $locales = explode(CRM_Core_DAO::VALUE_SEPARATOR, $domain->locales);
 
@@ -437,12 +453,27 @@ ADD UNIQUE INDEX `unique_entity_id` ( `entity_id` )";
 
       // now check for all fields if the index exists
       foreach ($fields as $field) {
-        // handle indices over substrings, CRM-6245
-        // $lengthName is appended to index name, $lengthSize is the field size modifier
-        $lengthName = isset($substrLenghts[$table][$field]) ? "_{$substrLenghts[$table][$field]}" : '';
-        $lengthSize = isset($substrLenghts[$table][$field]) ? "({$substrLenghts[$table][$field]})" : '';
+        $fieldName = implode('_', (array) $field);
 
-        $names = array("index_{$field}{$lengthName}", "FK_{$table}_{$field}{$lengthName}", "UI_{$field}{$lengthName}", "{$createIndexPrefix}_{$field}{$lengthName}");
+        if (is_array($field)) {
+          // No support for these for combined indexes as yet - add a test when you
+          // want to add that.
+          $lengthName = '';
+          $lengthSize = '';
+        }
+        else {
+          // handle indices over substrings, CRM-6245
+          // $lengthName is appended to index name, $lengthSize is the field size modifier
+          $lengthName = isset($substrLenghts[$table][$fieldName]) ? "_{$substrLenghts[$table][$fieldName]}" : '';
+          $lengthSize = isset($substrLenghts[$table][$fieldName]) ? "({$substrLenghts[$table][$fieldName]})" : '';
+        }
+
+        $names = array(
+          "index_{$fieldName}{$lengthName}",
+          "FK_{$table}_{$fieldName}{$lengthName}",
+          "UI_{$fieldName}{$lengthName}",
+          "{$createIndexPrefix}_{$fieldName}{$lengthName}",
+        );
 
         // skip to the next $field if one of the above $names exists; handle multilingual for CRM-4126
         foreach ($names as $name) {
@@ -454,40 +485,56 @@ ADD UNIQUE INDEX `unique_entity_id` ( `entity_id` )";
 
         // the index doesn't exist, so create it
         // if we're multilingual and the field is internationalised, do it for every locale
-        if (!CRM_Utils_System::isNull($locales) and isset($columns[$table][$field])) {
+        // @todo remove is_array check & add multilingual support for combined indexes and add a test.
+        // Note combined indexes currently using this function are on fields like
+        // entity_id + entity_table which are not multilingual.
+        if (!is_array($field) && !CRM_Utils_System::isNull($locales) and isset($columns[$table][$fieldName])) {
           foreach ($locales as $locale) {
-            $queries[] = "CREATE INDEX {$createIndexPrefix}_{$field}{$lengthName}_{$locale} ON {$table} ({$field}_{$locale}{$lengthSize})";
+            $queries[] = "CREATE INDEX {$createIndexPrefix}_{$fieldName}{$lengthName}_{$locale} ON {$table} ({$fieldName}_{$locale}{$lengthSize})";
           }
         }
         else {
-          $queries[] = "CREATE INDEX {$createIndexPrefix}_{$field}{$lengthName} ON {$table} ({$field}{$lengthSize})";
+          $queries[] = "CREATE INDEX {$createIndexPrefix}_{$fieldName}{$lengthName} ON {$table} (" . implode(',', (array) $field) . "{$lengthSize})";
         }
       }
     }
 
     // run the queries without i18n-rewriting
-    $dao = new CRM_Core_DAO;
+    $dao = new CRM_Core_DAO();
     foreach ($queries as $query) {
       $dao->query($query, FALSE);
     }
   }
 
   /**
-   * @param $customFieldID
-   * @param $tableName
-   * @param $columnName
+   * Drop an index if one by that name exists.
+   *
+   * @param string $tableName
+   * @param string $indexName
+   */
+  public static function dropIndexIfExists($tableName, $indexName) {
+    if (self::checkIfIndexExists($tableName, $indexName)) {
+      CRM_Core_DAO::executeQuery("DROP INDEX $indexName ON $tableName");
+    }
+  }
+
+  /**
+   * @param int $customFieldID
+   * @param string $tableName
+   * @param string $columnName
    * @param $length
    *
    * @throws Exception
    */
-  static function alterFieldLength($customFieldID, $tableName, $columnName, $length) {
+  public static function alterFieldLength($customFieldID, $tableName, $columnName, $length) {
     // first update the custom field tables
     $sql = "
 UPDATE civicrm_custom_field
 SET    text_length = %1
 WHERE  id = %2
 ";
-    $params = array(1 => array($length, 'Integer'),
+    $params = array(
+      1 => array($length, 'Integer'),
       2 => array($customFieldID, 'Integer'),
     );
     CRM_Core_DAO::executeQuery($sql, $params);
@@ -519,13 +566,82 @@ MODIFY      {$columnName} varchar( $length )
     }
     else {
       CRM_Core_Error::fatal(ts('Could Not Find Custom Field Details for %1, %2, %3',
-          array(
-            1 => $tableName,
-            2 => $columnName,
-            3 => $customFieldID
-          )
-        ));
+        array(
+          1 => $tableName,
+          2 => $columnName,
+          3 => $customFieldID,
+        )
+      ));
     }
   }
-}
 
+  /**
+   * Check if the table has an index matching the name.
+   *
+   * @param string $tableName
+   * @param array $indexName
+   *
+   * @return \CRM_Core_DAO|object
+   */
+  public static function checkIfIndexExists($tableName, $indexName) {
+    $result = CRM_Core_DAO::executeQuery(
+      "SHOW INDEX FROM $tableName WHERE key_name = %1 AND seq_in_index = 1",
+      array(1 => array($indexName, 'String'))
+    );
+    if ($result->fetch()) {
+      return TRUE;
+    }
+    return FALSE;
+  }
+
+  /**
+   * Check if the table has a specified column
+   *
+   * @param string $tableName
+   * @param string $columnName
+   *
+   * @return \CRM_Core_DAO|object
+   */
+  public static function checkIfFieldExists($tableName, $columnName) {
+    $result = CRM_Core_DAO::executeQuery(
+      "SHOW COLUMNS FROM $tableName LIKE %1",
+      array(1 => array($columnName, 'String'))
+    );
+    if ($result->fetch()) {
+      return TRUE;
+    }
+    return FALSE;
+  }
+
+  /**
+   * Remove a foreign key from a table if it exists
+   *
+   * @param $table_name
+   * @param $constraint_name
+   */
+  public static function safeRemoveFK($table_name, $constraint_name) {
+
+    $config = CRM_Core_Config::singleton();
+    $dbUf = DB::parseDSN($config->dsn);
+    $query = "
+      SELECT CONSTRAINT_NAME FROM INFORMATION_SCHEMA.TABLE_CONSTRAINTS
+      WHERE TABLE_SCHEMA = %1
+      AND TABLE_NAME = %2
+      AND CONSTRAINT_NAME = %3
+      AND CONSTRAINT_TYPE = 'FOREIGN KEY'
+    ";
+    $params = array(
+      1 => array($dbUf['database'], 'String'),
+      2 => array($table_name, 'String'),
+      3 => array($constraint_name, 'String'),
+    );
+    $dao = CRM_Core_DAO::executeQuery($query, $params);
+
+    if ($dao->fetch()) {
+      CRM_Core_DAO::executeQuery("ALTER TABLE {$table_name} DROP FOREIGN KEY {$constraint_name}", array());
+      return TRUE;
+    }
+    return FALSE;
+  }
+
+}
